requirements.txt
tinys3==0.1.11
futures==2.1.6
requests==2.5.1

Flask==0.10.1
Jinja2==2.7.1
Werkzeug==0.9.4
itsdangerous==0.23

ExifRead==2.1.1

Unidecode==0.04.18
python-slugify==1.1.2
